Refactor and move sendFile to its own module.
